% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{border_colours}
\alias{border_colours}
\title{Colour scheme for well borders}
\format{
A vector containing 4 colours
}
\usage{
border_colours
}
\description{
A colour scheme for the borders of wells that contains 4 colours. The colours are meant to
mark the status of a well (Fail, Check, OK and Not tested). They do not appear in the
\code{protti_colours} and \code{viridis_colours} schemes, which ensures that borders are visible on
every fill colour.
}
\keyword{datasets}
