% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_continuous_24}
\alias{data_continuous_24}
\title{Continuous 24-well plate dataset}
\format{
A data frame with a \code{Value}, a \code{well} and a \code{Status} column.
}
\source{
Randomly generated.
}
\usage{
data_continuous_24
}
\description{
A dataset containing 24 positive numeric values randomly generated using a normal distribution (\code{rnorm()}).
Each value is assigned to a position in a 24-well plate.
The \code{Status} column marks wells with a low signal.
}
\keyword{datasets}
