% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_continuous_384}
\alias{data_continuous_384}
\title{Continuous 384-well plate dataset}
\format{
A data frame with a \code{Value}, a \code{well} and a \code{Status} column.
}
\source{
Randomly generated.
}
\usage{
data_continuous_384
}
\description{
A dataset containing 384 positive numeric values randomly generated using a normal distribution (\code{rnorm()}).
Each value is assigned to a position in a 384-well plate.
The \code{Status} column marks the wells at the edge of the plate, which are prone to evaporation.
}
\keyword{datasets}
