% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_continuous_48}
\alias{data_continuous_48}
\title{Continuous 48-well plate dataset}
\format{
A data frame with a \code{Value}, a \code{well} and a \code{Status} column.
}
\source{
Randomly generated.
}
\usage{
data_continuous_48
}
\description{
A dataset containing 48 positive numeric values randomly generated using a normal distribution (\code{rnorm()}).
Each value is assigned to a position in a 48-well plate.
The \code{Status} column marks wells with a low or a high signal.
}
\keyword{datasets}
