% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_continuous_6}
\alias{data_continuous_6}
\title{Continuous 6-well plate dataset}
\format{
A data frame with a \code{Value}, a \code{well} and a \code{Status} column.
}
\source{
Randomly generated.
}
\usage{
data_continuous_6
}
\description{
A dataset containing 6 positive numeric values randomly generated using a normal distribution (\code{rnorm()}).
Each value is assigned to a position in a 6-well plate.
The \code{Status} column marks the well with a very low signal.
}
\keyword{datasets}
