% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_continuous_96}
\alias{data_continuous_96}
\title{Continuous 96-well plate dataset}
\format{
A data frame with a \code{Value}, a \code{well} and a \code{Status} column.
}
\source{
Randomly generated.
}
\usage{
data_continuous_96
}
\description{
A dataset containing 96 positive numeric values randomly generated using a normal distribution (\code{rnorm()}).
Each value is assigned to a position in a 96-well plate.
The \code{Status} column marks wells with a low signal.
}
\keyword{datasets}
