% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_discrete_96}
\alias{data_discrete_96}
\title{Discrete 96-well plate dataset}
\format{
A data frame with a \code{Compound}, a \code{well}, a \code{Compound_multiline} and a \code{Replicate} column.
}
\source{
Drugs were chosen from a standard FDA approved drug library.
}
\usage{
data_discrete_96
}
\description{
A dataset containing 22 conditions for a hypothetical drug treatment in a 96-well plate.
The first column contains the positive control while the last column contains the negative control.
Each treatment is assigned to four wells in a 96-well plate.
The \code{Replicate} column contains the number of the replicate of each compound. The two controls
have eight replicates, every other compound has four.
}
\keyword{datasets}
