% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viridis_colours}
\alias{viridis_colours}
\title{Viridis colour scheme}
\format{
A vector containing 256 colours
}
\source{
viridis R package
}
\usage{
viridis_colours
}
\description{
A colour scheme by the viridis colour scheme from the viridis R package.
}
\keyword{datasets}
