% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_discrete_96}
\alias{data_discrete_96}
\title{Discrete 96-well plate dataset}
\format{
A data frame with a \code{Value} and a \code{well} column.
}
\source{
Drugs were chosen from a standard FDA approved drug library.
}
\usage{
data_discrete_96
}
\description{
A dataset containing 22 conditions for a hypothetical drug treatment in a 96-well plate.
The first column contains the negative control while the last column contains the positive control.
Each treatment is assigned to four wells in a 96-well plate.
}
\keyword{datasets}
