% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protti_colours}
\alias{protti_colours}
\title{Colour scheme for protti}
\format{
A vector containing 100 colours
}
\source{
protti R package.
}
\usage{
protti_colours
}
\description{
A colour scheme from the \code{protti} package that contains 100 colours.
}
\keyword{datasets}
